requirements.txt
portia-sdk-python[mistral]
python-dotenv
rich  # For better terminal output 